Modificaciones Ejercicios entrada y salida

+++ ejercicios/01_entrada_salida/ejercicio_10/app.py
@@ -42,8 +42,8 @@ class App(customtkinter.CTk):
     def btn_mostrar_on_click(self):
       importe_txt = self.txt_importe.get()
       descuento_txt = self.txt_descuento.get()
-      importe_num = float (importe_txt)
-      descuento_num = float (descuento_txt)
+      importe_num = int (importe_txt)
+      descuento_num = int (descuento_txt)
       descuento_total = importe_num * descuento_num / 100
       importe_act_num = importe_num - descuento_total
       mensaje = "El importe final con el descuento aplicado es de: {0}".format(importe_act_num)
